Daniel's Ex 7
This is the fourth API I've been working on. The first one, rats, was
difficult to work with as the function I would try to make to count
words would not work as intended, so I tried moving on to another API
where I encountered the same problem and a weather.json file I had saved
locally but would still bring up errors. This API I've used is from the
internet and takes weather data from a Cairns, Australia. Te sketch
displays objects based on provided weather data.